Missingness Map
Leveraging the Amelia package, missingness map is a great way to get a quick visualization of your missing information based on each column. If you have factorized a column with blank values it will not show up on the missingness map. A recommendation is to make this your first process for data pre-processing.